further clarification
well.. to be honest, I have never understood or tried to understand
what happened. It’s simply too annoying and wasting of time. To me,
writing guides and tools could benefit a lot of people and save their
human labor (or even money, if they ever paid). Of course there are
other guide writers, but what’s the point of compete, or even attacking
each other?